+++ OpenBox.h
@@ -1,0 +1,38 @@
+
+#ifndef OPENBOX_H
+#define OPENBOX_H
+
+#include "SecureBox.h"
+
+//================================================================================
+// Function: solveWithGauss
+// Description: Решает систему линейных уравнений по модулю 2
+//              методом Гаусса. Используется для вычисления, какие
+//              кнопки нужно нажать для открытия коробки.
+//================================================================================
+bool solveWithGauss(
+    std::vector<std::vector<int>>& matrix, // матриця впливу (A) — NxN
+    std::vector<int>& targetState, // вектор початкового стану (b)
+    std::vector<int>& solution // Вектор рішення: які кнопки натискати (x)
+);
+
+//================================================================================
+// Function: createEquationSystem
+// Description: Строит матрицу влияния всех кнопок и вектор
+//              текущего состояния ячеек (включено/выключено).
+//================================================================================
+void createEquationSystem(
+    const std::vector<std::vector<bool>>& boxState,
+    std::vector<std::vector<int>>& influenceMatrix,
+    std::vector<int>& initialStateVector
+);
+//================================================================================
+// Function: openBox
+// Description: Основная функция, открывающая коробку. Строит уравнение,
+//              решает его, применяет найденные нажатия и возвращает,
+//              осталась ли коробка заблокированной.
+//================================================================================
+bool openBox(uint32_t rows, uint32_t cols);
+
+
+#endif //OPENBOX_H
