+++ SecureBox.h
@@ -1,0 +1,104 @@
+#ifndef SECUREBOX_H
+#define SECUREBOX_H
+#include <vector>
+#include <random>
+#include <time.h>
+#include <iostream>
+
+/*
+You are given a locked container represented as a two-dimensional grid of boolean values (true = locked, false = unlocked).
+Your task is to write an algorithm that fully unlocks the box, i.e.,
+transforms the entire matrix into all false.
+
+Implement the function:
+bool openBox(uint32_t y, uint32_t x);
+This function should:
+    - Use the SecureBox public API (toggle, isLocked, getState).
+    - Strategically toggle cells to reach a state where all elements are false.
+    - Return true if the box remains locked, false if successfully unlocked.
+You are not allowed to path or modify the SecureBox class.
+
+Evaluation Criteria:
+    - Functional correctness
+    - Computational efficiency
+    - Code quality, structure, and comments
+    - Algorithmic insight and clarity
+*/
+
+class SecureBox
+{
+private:
+    std::vector<std::vector<bool>> box;
+
+public:
+
+    //================================================================================
+    // Constructor: SecureBox
+    // Description: Initializes the secure box with a given size and
+    //              shuffles its state using a pseudo-random number generator
+    //              seeded with current time.
+    //================================================================================
+    SecureBox(uint32_t y, uint32_t x): ySize(y), xSize(x)
+    {
+        rng.seed(time(0));
+        box.resize(y);
+        for (auto& it : box)
+            it.resize(x);
+        shuffle();
+    }
+
+    //================================================================================
+    // Method: toggle
+    // Description: Toggles the state at position (x, y) and also all cells in the
+    //              same row above and the same column to the left of it.
+    //================================================================================
+    void toggle(uint32_t y, uint32_t x)
+    {
+        box[y][x] = !box[y][x];
+        for (uint32_t i = 0; i < xSize; i++)
+            box[y][i] = !box[y][i];
+        for (uint32_t i = 0; i < ySize; i++)
+            box[i][x] = !box[i][x];
+    }
+
+    //================================================================================
+    // Method: isLocked
+    // Description: Returns true if any cell
+    //              in the box is true (locked); false otherwise.
+    //================================================================================
+    bool isLocked()
+    {
+        for (uint32_t x = 0; x < xSize; x++)
+            for (uint32_t y = 0; y < ySize; y++)
+                if (box[y][x])
+                    return true;
+
+        return false;
+    }
+
+    //================================================================================
+    // Method: getState
+    // Description: Returns a copy of the current state of the box.
+    //================================================================================
+    std::vector<std::vector<bool>> getState()
+    {
+        return box;
+    }
+
+private:
+    std::mt19937_64 rng;
+    uint32_t ySize, xSize;
+
+    //================================================================================
+    // Method: shuffle
+    // Description: Randomly toggles cells in the box to
+    // create an initial locked state.
+    //================================================================================
+    void shuffle()
+    {
+        for (uint32_t t = rng() % 1000; t > 0; t--)
+            toggle(rng() % ySize, rng() % xSize);
+    }
+};
+
+#endif //SECUREBOX_H
